requirements.txt
discord.py==2.3.2
yt-dlp==2025.04.16
pynacl==1.5.0